pyproject.toml
[tool.poetry]
name = "agent-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.25"
langchain-core = "^0.3.59"
langchain-community = "^0.3.24"
langchain-openai = "^0.3.16"
python-dotenv = "^1.0.1"
streamlit = "^1.37.0"
qdrant-client = "^1.14.2"
langgraph = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
